+++ src/deadlink/__init__.py
@@ -1,5 +1,4 @@
 from . import _cli
-from .__about__ import __version__
 from ._main import categorize_urls
 
-__all__ = ["__version__", "categorize_urls", "_cli"]
+__all__ = ["categorize_urls", "_cli"]